imageupload/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
Pillow==9.4.0
pydantic==1.10.4
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
